# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Swiftdrop Delivery API
  version: 1.0.0
  description: >
    Multi-tenant last-mile delivery management platform.


    Each **company** is an isolated tenant. An **ADMIN** owns a company and
    manages its delivery agents and parcels. A **DELIVERY_AGENT** belongs to one
    company and handles the deliveries assigned to them. A **SUPER_ADMIN** is a
    platform operator who monitors and manages every company.


    ## Authentication

    Every endpoint (except the ones tagged as public) requires **two** headers:

    - `x-api-key`: a shared API key (`API_KEY` env var, default `dev-api-key`).
    - `Authorization: Bearer <JWT>`: obtained from `POST /auth/login` or
      `POST /auth/register`.


    ## Roles

    | Role            | Scope                                                        |
    |-----------------|-------------------------------------------------------------|
    | `SUPER_ADMIN`   | Cross-tenant: monitor everything, create companies/agents.  |
    | `ADMIN`         | Owns one company: manage its agents, parcels, reports.      |
    | `DELIVERY_AGENT`| Belongs to one company: handle assigned parcels.            |


    The role required for each operation is stated in the operation summary,
    e.g. *(Admin)*. Operations marked *(Any authenticated)* accept any logged-in
    user, but the service layer still scopes data to the caller's company.

  contact:
    name: Swiftdrop API Support
servers:
  - url: http://localhost:3000
...
security:
  - ApiKeyAuth: []
    BearerAuth: []
paths:
...
  /platform/overview:
    get:
      tags: [Platform (Super Admin)]
      operationId: getPlatformOverview
      summary: Platform overview (Super Admin)
      description: Cross-tenant totals plus per-company statistics.
      responses:
        '200':
          description: Aggregated platform statistics.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlatformOverview'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /platform/companies:
    get:
      tags: [Platform (Super Admin)]
      operationId: listPlatformCompanies
      summary: List all companies (Super Admin)
      description: Every company on the platform with owner and volume counts.
      responses:
        '200':
          description: Companies across all tenants.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlatformCompany'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Platform (Super Admin)]
      operationId: createPlatformCompany
      summary: Create a company + owner admin (Super Admin)
      description: Provisions a new tenant and its first ADMIN in one transaction.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompanyDto'
      responses:
        '201':
          description: Company and owner admin created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCompanyResult'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /platform/users:
    get:
      tags: [Platform (Super Admin)]
      operationId: listPlatformUsers
      summary: List all users (Super Admin)
      description: Every admin and delivery agent across all companies.
      responses:
        '200':
          description: Users across all tenants.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlatformUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /platform/agents:
    post:
      tags: [Platform (Super Admin)]
      operationId: createPlatformAgent
      summary: Create a delivery agent for any company (Super Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlatformAgentDto'
      responses:
        '201':
          description: Delivery agent created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /platform/parcels:
    get:
      tags: [Platform (Super Admin)]
      operationId: listPlatformParcels
      summary: List all parcels (Super Admin)
      description: Live view of every parcel across all companies.
      responses:
        '200':
          description: Parcels across all tenants.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlatformParcel'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /companies:
    get:
      tags: [Companies]
      operationId: listCompanies
      summary: List companies (Any authenticated)
      description: >
        Lightweight list of companies (id, name, join code). Used by agents to
        pick a company to switch into.
      responses:
        '200':
          description: Company summaries.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Company'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /users:
    post:
      tags: [Users]
      operationId: createDeliveryAgent
      summary: Create a delivery agent (Admin)
      description: Creates a delivery agent within the admin's company.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeliveryAgentDto'
      responses:
        '201':
          description: Delivery agent created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      tags: [Users]
      operationId: listUsers
      summary: List users in my company (Admin)
      responses:
        '200':
          description: Users belonging to the admin's company.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /parcels:
    post:
      tags: [Parcels]
      operationId: createParcel
      summary: Register a parcel (Admin)
      description: >
        Creates a parcel in the admin's company. `trackingNumber` is generated
        when omitted.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateParcelDto'
      responses:
        '201':
          description: Parcel registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Parcel'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    get:
      tags: [Parcels]
      operationId: listParcels
      summary: List parcels (Any authenticated)
      description: >
        Admins see all parcels in their company; delivery agents see only the
        parcels assigned to them.
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ParcelStatus'
        - name: sender
          in: query
          required: false
          description: Case-insensitive partial match on sender name (admins).
          schema:
            type: string
      responses:
        '200':
          description: Parcels visible to the caller.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Parcel'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Shared API key required on every non-public request.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by /auth/login or /auth/register.
...
  responses:
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            statusCode: 400
            message:
              - email must be an email
            error: Bad Request
    Unauthorized:
      description: Missing/invalid API key or JWT.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            statusCode: 401
            message: Unauthorized
            error: Unauthorized
    Forbidden:
      description: Authenticated but not allowed for this role/tenant.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            statusCode: 403
            message: Insufficient permissions
            error: Forbidden
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            statusCode: 404
            message: Resource not found
            error: Not Found
    Conflict:
      description: Resource already exists (e.g. duplicate email).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            statusCode: 409
            message: A user with this email already exists
            error: Conflict
  schemas:
    UserRole:
      type: string
      enum: [SUPER_ADMIN, ADMIN, DELIVERY_AGENT]
    ParcelStatus:
      type: string
      enum: [REGISTERED, PICKED_UP, OUT_FOR_DELIVERY, DELIVERED, FAILED_ATTEMPT]
...
    Error:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        error:
          type: string
          example: Bad Request
        timestamp:
          type: string
          format: date-time
        path:
          type: string
      required: [statusCode, message]
...
    Company:
      type: object
      description: Lightweight company summary.
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        joinCode:
          type: string
          example: SWIFT1
    CompanyDetail:
      allOf:
        - $ref: '#/components/schemas/Company'
        - type: object
          properties:
            ownerId:
              type: string
              format: uuid
              nullable: true
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    CreateDeliveryAgentDto:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 1
          example: Jane Smith
        email:
          type: string
          format: email
          example: jane@swiftdrop.com
        password:
          type: string
          format: password
          minLength: 6
          example: password123
        isAvailable:
          type: boolean
          default: true
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        companyId:
          type: string
          format: uuid
          nullable: true
        isAvailable:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateParcelDto:
      type: object
      required: [senderName, senderAddress, recipientName, recipientAddress]
      properties:
        trackingNumber:
          type: string
          maxLength: 50
          description: Auto-generated when omitted.
          example: SWFT-2K5J8N
        senderName:
          type: string
          maxLength: 255
          example: Warehouse A
        senderAddress:
          type: string
          example: 12 Depot Road
        recipientName:
          type: string
          maxLength: 255
          example: John Doe
        recipientAddress:
          type: string
          example: 88 Maple Street
...
    Parcel:
      type: object
      properties:
        id:
          type: string
          format: uuid
        trackingNumber:
          type: string
        companyId:
          type: string
          format: uuid
        senderName:
          type: string
        senderAddress:
          type: string
        recipientName:
          type: string
        recipientAddress:
          type: string
        assignedAgentId:
          type: string
          format: uuid
          nullable: true
        status:
          $ref: '#/components/schemas/ParcelStatus'
        retryQueued:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateCompanyDto:
      type: object
      required: [companyName, adminName, adminEmail, adminPassword]
      properties:
        companyName:
          type: string
          minLength: 1
          example: Metro Couriers
        adminName:
          type: string
          minLength: 1
          example: Bob Admin
        adminEmail:
          type: string
          format: email
          example: bob@metro.com
        adminPassword:
          type: string
          format: password
          minLength: 6
          example: password123
    CreatePlatformAgentDto:
      type: object
      required: [companyId, name, email, password]
      properties:
        companyId:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 6
        isAvailable:
          type: boolean
          default: true
    CreatedCompanyAdmin:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        companyId:
          type: string
          format: uuid
          nullable: true
    CreateCompanyResult:
      type: object
      properties:
        company:
          $ref: '#/components/schemas/CompanyDetail'
        admin:
          $ref: '#/components/schemas/CreatedCompanyAdmin'
    PlatformCompany:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        joinCode:
          type: string
        ownerName:
          type: string
          nullable: true
        ownerEmail:
          type: string
          nullable: true
        agentCount:
          type: integer
        parcelCount:
          type: integer
        deliveredCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    PlatformUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        companyId:
          type: string
          format: uuid
          nullable: true
        companyName:
          type: string
          nullable: true
        isAvailable:
          type: boolean
        createdAt:
          type: string
          format: date-time
    PlatformParcel:
      type: object
      properties:
        id:
          type: string
          format: uuid
        trackingNumber:
          type: string
        status:
          $ref: '#/components/schemas/ParcelStatus'
        retryQueued:
          type: boolean
        companyId:
          type: string
          format: uuid
        companyName:
          type: string
          nullable: true
        assignedAgentId:
          type: string
          format: uuid
          nullable: true
        assignedAgentName:
          type: string
          nullable: true
        senderName:
          type: string
        recipientName:
          type: string
        createdAt:
          type: string
          format: date-time
    PlatformOverview:
      type: object
      properties:
        totals:
          type: object
          properties:
            companies:
              type: integer
            admins:
              type: integer
            agents:
              type: integer
            parcels:
              type: integer
            parcelsByStatus:
              type: object
              additionalProperties:
                type: integer
              example:
                REGISTERED: 4
                OUT_FOR_DELIVERY: 2
                DELIVERED: 10
        companies:
          type: array
          items:
            $ref: '#/components/schemas/PlatformCompany'